+++ tests/prop.test.js
@@ -13,4 +13,16 @@ describe("Props", () => {
         expect(col.textContent).toBe("DIV");
     });
 
+    test("$refs", () => {
+        document.body.innerHTML = `<div id="app">
+            <div :ref="title">Title Head</div>
+            <div id="col" :text="() => $refs.title.textContent"></div>
+        </div>`;
+
+        let lf = new leaf("#app");
+        let col = document.querySelector("#col");
+
+        expect(col.textContent).toBe("Title Head");
+    });
+
 });